Add a first framework for vminspect
The current state is a rudimentary framework for vminspect. In its
current form, it parses a Rekall profile and prepares it in a way, such
that it can be further processed for injecting tap points into the guest
kernel. Some of these system calls are Also, it initializes Xen and
LibVMI for further processing.

Please note that the Rekall profile for ARM comprises ambiguous
system calls (i.e., we have multiple system calls that map to the same
address). Some of these system calls are functions for 16 bit
functionality. Yet, they are part of the profile.

The current state does not (yet) inject any tap points or call backs
that are invoked by Xen and LibVMI on various guest events.

Besides, the current build is not (yet) adjusted for Yocto. For now, we
use a normal Makefile to build vminspect.

Signed-off-by: Sergej Proskurin <proskurin@sec.in.tum.de>