Adjusted policy handling code.
Now accomodate event that "triggers" and "untriggers", e.g. process death. It "triggers" because the BP must be in the trigger view, and it untriggers because it indicates a process is dying. We may run into problems if we apply this logic to a process from which we see no events except its death.